requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
python-dotenv
requests
asyncpg